Set link args in build script
The primary purpose of `config.toml` was to set the default target and
required subsystem link args for drivers. The problem with its use is
that it is a project-level configuration for cargo, and not a
package-level configuration. It will not work in the case of:

- package directories in a workspace
- package directory when `--manifest-path` is used

Use `build.rs` instead to provide the link args, which will be executed
before a package is built.

Ref: https://doc.rust-lang.org/cargo/reference/config.html#hierarchical-structure
Ref: https://doc.rust-lang.org/cargo/reference/build-scripts.html
Signed-off-by: Tim Crawford <tcrawford@system76.com>